Initial TicTacToe Build
Added an app drawn board instead of an image just for tutorial following
purposes.

Pieces are just for show thus far.

A few functions already included, although not yet tested:

renderTile(): where tiles are decided and rendered
componentDiDMount(): tests if game initialize or not
initializeGame(): where the board pieces are saved (I will rename it
later, just realized that the name isn't good on this one)