Big changes
We are ditching React Native for this project because we are pretty much learning Javascript, with the challenge in time, and schedules, learning it and React Native proved a challenge. Thus, we are ditching it to make the new version of the app in Javascript and some markup language, probably HTML+CSS, or some variant like Pug and Stylus.

startGame(): calls the initialization of the game (also for replays), initializes the board, connecting the cells with the CSS and adds a click function to look for them. 

function turnClick(): will be adding the players function.

function turn(): will be adding turn controls.