Huge Refactor - Please Read
Since we still needed to add a way to get usernames, a refactor on how
messages were displayed and the board was done was needed.

Before, just the variable document.turn would do the job. Now, there are
more variables controlling it, and a bunch of changes.

The main one is that one of the main methods, turnControl(), was split
in two. It became the observer() (as in the Observer Pattern, because
that's what it is), and the resultDisplay(), which makes a little more
elegant, SOLID principle, and distributes functionality better, besides
solving the problem of the single variable for board and messages.

I'm reopening the issue with Draws because I didn't adjust to display a
drawing message with the refactor yet.